Create HW11_GuessingGame.asm
Implementation of a guessing game on the PIC16F84A microcontroller in which you must guess which LED is on at any given time in an LED chaser sequence. This code includes such a chaser sequence, guessing functionality, what to do if an error occurs, and what to do once you win.